requirements.txt
tensorflow
flask
numpy
opencv-python
opencv-python-headless
Pillow
gunicorn
ortools
opencv-contrib-python
